feat(network-util-class): check if network connection is available
- structure application structure by feature
- check if the is network/slow network

[Finishes: #156187450]
