+++ app/build.gradle
@@ -36,19 +36,19 @@ dependencies {
     androidTestImplementation 'com.android.support.test.espresso:espresso-web:3.0.2'
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'de.hdodenhof:circleimageview:2.2.0'
-    implementation 'com.squareup.picasso:picasso:2.71828'
-    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
-    implementation 'com.jakewharton.picasso:picasso2-okhttp3-downloader:1.1.0'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:support-annotations:27.1.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-contrib:3.0.2'
-    implementation 'com.android.support.test.espresso:espresso-idling-resource:3.0.2'
-    implementation 'com.android.support:exifinterface:27.1.1'
+    implementation "com.android.support:appcompat-v7:${rootProject.supportVersion}"
+    implementation "com.android.support.constraint:constraint-layout:${rootProject.constraintLayoutVersion}"
+    implementation "com.android.support:recyclerview-v7:${rootProject.recyclerViewVersion}"
+    implementation "com.android.support:cardview-v7:${rootProject.cardViewVersion}"
+    implementation "de.hdodenhof:circleimageview:${rootProject.circleImageVersion}"
+    implementation "com.squareup.picasso:picasso:${rootProject.picassoVersion}"
+    implementation "com.squareup.retrofit2:retrofit:${rootProject.retrofitVersion}"
+    implementation "com.squareup.retrofit2:converter-gson:${rootProject.converterGsonVersion}"
+    implementation "com.jakewharton.picasso:picasso2-okhttp3-downloader:${rootProject.okHTTPVersion}"
+    implementation "com.android.support:design:${rootProject.designVersion}"
+    implementation "com.android.support:support-annotations:${rootProject.annotationsVersion}"
+    androidTestImplementation "com.android.support.test.espresso:espresso-core:${rootProject.espressoCoreVersion}"
+    androidTestImplementation "com.android.support.test.espresso:espresso-contrib:${rootProject.espressoContribVersion}"
+    implementation "com.android.support.test.espresso:espresso-idling-resource:${rootProject.espressoIdlingResourceVersion}"
+    implementation "com.android.support:exifinterface:${rootProject.exifInterfaceVersion}"
 }
