+++ app/src/main/java/com/andela/omokolataiwo/levelup/home/models/GithubProfile.java
@@ -1,4 +1,4 @@
-package com.andela.omokolataiwo.levelup.models;
+package com.andela.omokolataiwo.levelup.home.models;
 
 import android.os.Parcel;
 import android.os.Parcelable;
