+++ app/src/main/java/com/andela/omokolataiwo/levelup/service/GithubAPI.java
@@ -3,7 +3,7 @@ package com.andela.omokolataiwo.levelup.service;
 import retrofit2.Call;
 import retrofit2.http.GET;
 
-import com.andela.omokolataiwo.levelup.models.GithubProfileResponse;
+import com.andela.omokolataiwo.levelup.home.models.GithubProfileResponse;
 
 /**
  * Github request API.
